garmin-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
garth==0.4.46
garminconnect==0.2.19
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
cryptography==41.0.7
email-validator==2.1.0
